+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Added
+* Added support for material texture assignments for wimdo exports.
+
 ### Changed
 * Changed the import behavior for wismhd maps to use collection instances instead of linked duplicate meshes.